Update for bare metal support
using parking lot prevents running in a fully embedded system (no OS) so
we replaced this in the implementation and created a new compilation
target for baremetal.

Fixing transient failure in test

Root cause: generate_domain_id() uses a global static AtomicUsize counter. Tests run
in parallel within the same process, so whichever test creates a domain first advances
the counter. The old test pre-registered CHILD_ID = 1 with the platform and then
asserted the generated ID would also be 1 — a race.

Fix: Create the capabilities first to get the real generated ID, then register it with
the platform and set the core domain using that actual ID. The hardcoded CHILD_ID = 1
constant is gone; the test now works regardless of what value the global counter is
at.

Fixing compilation warning

Unused variable.

+++ 2026/src/domain.rs
@@ -5,14 +5,14 @@ use crate::capability::{CapabilityWeak, LocalHandle};
 use crate::memory::MemoryRegion;
 use alloc::collections::BTreeMap;
 use alloc::vec::Vec;
-use core::sync::atomic::{AtomicU64, Ordering};
+use core::sync::atomic::{AtomicUsize, Ordering};
 
 /// Global domain ID counter
-static NEXT_DOMAIN_ID: AtomicU64 = AtomicU64::new(1);
+static NEXT_DOMAIN_ID: AtomicUsize = AtomicUsize::new(1);
 
 /// Generate a unique domain ID
 pub fn generate_domain_id() -> u64 {
-    NEXT_DOMAIN_ID.fetch_add(1, Ordering::SeqCst)
+    NEXT_DOMAIN_ID.fetch_add(1, Ordering::SeqCst) as u64
 }
 
 /// Domain status
